feat!: implement support for parent sites required consents
BREAKING CHANGE: Before this change, the library did not handle situation where the project consent schemas had a set of required consents, that may be different from the global parent site this current project is in (on CDC side).
As well, the library can now automate the proces s of declining optional consents that require a default value in order to complete registration. This is done on top of the regular acceptance of mandatory project consents.